+++ lib/middleware.js
@@ -5,16 +5,13 @@ const path = require('path');
 const mime = require('mime');
 
 const DevMiddlewareError = require('./DevMiddlewareError');
-const {
-  getFilenameFromUrl,
-  handleRangeHeaders,
-  handleRequest,
-  ready,
-} = require('./utils');
+const { getFilenameFromUrl, handleRangeHeaders, ready } = require('./utils');
 
 // Do not add a charset to the Content-Type header of these file types
 // otherwise the client will fail to render them correctly.
 const NonCharsetFileTypes = /\.(wasm|usdz)$/;
+
+const HASH_REGEXP = /[0-9a-f]{10,}/;
 
 module.exports = function wrapper(context) {
   return function middleware(req, res, next) {
@@ -61,7 +58,6 @@ module.exports = function wrapper(context) {
     }
 
     return new Promise((resolve) => {
-      handleRequest(context, filename, processRequest, req);
       // eslint-disable-next-line consistent-return
       function processRequest() {
         try {
@@ -131,6 +127,27 @@ module.exports = function wrapper(context) {
 
         resolve();
       }
+
+      if (
+        context.options.lazy &&
+        (!context.options.filename || context.options.filename.test(filename))
+      ) {
+        context.rebuild();
+      }
+
+      if (HASH_REGEXP.test(filename)) {
+        try {
+          if (context.fs.statSync(filename).isFile()) {
+            processRequest();
+
+            return;
+          }
+        } catch (_error) {
+          // Ignore error
+        }
+      }
+
+      ready(context, processRequest, req);
     });
   };
 };
