+++ lib/utils/getFilenameFromUrl.js
@@ -8,6 +8,7 @@ const querystring = require('querystring');
 // see: https://github.com/webpack/webpack-dev-server/issues/641
 function getPaths(publicPath, compiler, url) {
   const compilers = compiler && compiler.compilers;
+
   if (Array.isArray(compilers)) {
     let compilerPublicPath;
 
@@ -37,6 +38,7 @@ function getPaths(publicPath, compiler, url) {
       }
     }
   }
+
   return {
     publicPath,
     outputPath: compiler.outputPath,
