+++ lib/utils/index.js
@@ -1,15 +1,21 @@
 const getFilenameFromUrl = require('./getFilenameFromUrl');
 const handleRangeHeaders = require('./handleRangeHeaders');
-const handleRequest = require('./handleRequest');
 const ready = require('./ready');
+const reporter = require('./reporter');
+const setupHooks = require('./setupHooks');
+const setupLogger = require('./setupLogger');
 const setupOutputFileSystem = require('./setupOutputFileSystem');
+const setupRebuild = require('./setupRebuild');
 const setupWriteToDisk = require('./setupWriteToDisk');
 
 module.exports = {
   getFilenameFromUrl,
   handleRangeHeaders,
-  handleRequest,
   ready,
+  reporter,
+  setupHooks,
+  setupLogger,
   setupOutputFileSystem,
+  setupRebuild,
   setupWriteToDisk,
 };
