+++ lib/utils/reporter.js
